Creata a base version of run_pipeline and fix tests.
